\name{as_sp}
\alias{as_sp}
\title{Convert osmar object to sp object}
\usage{
  as_sp(obj, ...)
}
\arguments{
  \item{obj}{An \code{\link{osmar}} object}

  \item{...}{Ignored}
}
\value{
  A \link[sp]{sp} object
}
\description{
  Convert an osmar object to a \link[sp]{sp} object.
}
\note{
  Not yet implemented!
}

