\name{osmsource_api}
\alias{osmsource_api}
\title{API OSM data source}
\usage{
  osmsource_api(url =
  "http://api.openstreetmap.org/api/0.6/")
}
\arguments{
  \item{url}{URL of the API}
}
\description{
  OSM API version 0.6 data source; this source just
  supports to the basic request elements (see
  \code{\link{get_osm_elements}}).
}
\seealso{
  \code{\link{get_osm}}, \code{\link{get_osm_elements}}

  Other osmsource: \code{\link{osmsource_osmosis}}
}

